[C++] [PR84231] overload on cond_expr in template
A non-type-dependent COND_EXPR within a template is reconstructed with
the original operands, after one with non-dependent proxies is built
to determine its result type.  This is problematic because the
operands of a COND_EXPR determined to be an rvalue may have been
converted to denote their rvalue nature.  The reconstructed one,
however, won't have such conversions, so lvalue_kind may not recognize
it as an rvalue, which may lead to e.g. incorrect overload resolution
decisions.

If we mistake the COND_EXPR for as an lvalue, overload resolution
might regard a conversion sequence that binds it to a non-const
reference as viable, and then select that over one that binds it to a
const reference.  Only after template substitution would we rebuild
the COND_EXPR, realize it is an rvalue, and conclude the reference
binding is ill-formed, but at that point we'd have long discarded any
alternate candidates we could have used.

This patch verifies that, if a non-type-dependent COND_EXPRs is
recognized as an rvalues, so is the to-be-template-substituted one
created in its stead, so that overload resolution selects the correct
alternative.


for  gcc/cp/ChangeLog

	PR c++/84231
	* typeck.c (build_x_conditional_expr): Make sure the
	to-be-tsubsted expr is an rvalue when it should.

for  gcc/testsuite/g++.dg/ChangeLog

	PR c++/84231
	* pr84231.C: New.